refactor: extract middleware pipeline runner
